Mpl3115A2 basic implimentated. Removed unused code

+++ sources/devices/i2c_device/i2c_device.cpp
@@ -34,7 +34,7 @@ void I2cDevice::stop()
     close(m_file);
 }
 
-int I2cDevice::i2cAddress() const
+uint8_t I2cDevice::i2cAddress() const
 {
     return 0x0;
 }
